// src/kl-scraper-options/types/enum/kl-scraper-options-env-variable.enum.ts
export enum KLScraperOptionsEnvVariable {
  KL_SCRAPER_BROWSER_REQUEST_TIMEOUT = 'KL_SCRAPER_BROWSER_REQUEST_TIMEOUT',
  KL_SCRAPER_BROWSER_HEADLESS = 'KL_SCRAPER_BROWSER_HEADLESS',
  KL_SCRAPER_BROWSER_DELAY_BETWEEN_REQUESTS = 'KL_SCRAPER_BROWSER_DELAY_BETWEEN_REQUESTS',
  KL_SCRAPER_LOG_LEVEL = 'KL_SCRAPER_LOG_LEVEL',
}